Remove chemcialLog.notes (See commit message notes)
There is currently no need to have chemcial log notes.
This commit removes chemicalLog.notes from the UI
but has no change on the backend. This is incase I
want to re-implement this feature in the future. I can
simply reference this commit and undo these changes to
bring back chemicalLog.notes in the UI.